refactor: replace getMissingData with checkEmptyBranch and move action order
