+++ src/proxy/processors/push-action/index.ts
@@ -14,7 +14,7 @@ import { exec as checkCommitMessages } from './checkCommitMessages';
 import { exec as checkAuthorEmails } from './checkAuthorEmails';
 import { exec as checkUserPushPermission } from './checkUserPushPermission';
 import { exec as clearBareClone } from './clearBareClone';
-import { exec as getMissingData } from './getMissingData';
+import { exec as checkEmptyBranch } from './checkEmptyBranch';
 
 export {
   parsePush,
@@ -33,5 +33,5 @@ export {
   checkAuthorEmails,
   checkUserPushPermission,
   clearBareClone,
-  getMissingData,
+  checkEmptyBranch,
 };
